+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Fixed `slack` on `x86_64-darwin`
+- Use `x86_64-darwin` on `psi`
 - Use 1Password SSH agent on `sigma`
 - Removed `apollo`
 - Use `tailscale` on personal machines